+++ packages/gitmoji-changelog-cli/src/index.js
@@ -9,6 +9,12 @@ const { main } = require('./cli')
 logger.start(`gitmoji-changelog v${version}`)
 
 const args = yargs
+  .option('mode', {
+    alias: 'm',
+    default: 'init',
+    description: 'build from scratch or incremental changelog',
+    choices: ['init', 'incremental'],
+  })
   .option('format', {
     alias: 'f',
     default: 'markdown',